requirements.txt
# requirements.txt
Flask==2.3.3
python-docx==1.1.0
pypdf==3.17.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
werkzeug==3.1.1
sentence-transformers==2.2.2
torch==2.2.1

# >>> Para o banco de dados
Flask-SQLAlchemy==3.1.1       # integração Flask + SQLAlchemy
SQLAlchemy==2.0.23            # ORM
psycopg2-binary==2.9.9        # driver Postgres (necessário em produção)
google-generativeai==0.8.5    # para usar a API Gemini

